writing: create basic chapter structure
The writing/ directory will contain all of the "inputs" to the
book-building tools. This will include markdown files for the book
contents as well as any images or other data that is fed into the
automated typesetting tools.

For now, populate two fake chapters, each with an introduction in a
README.md and two sections. These will be organized later using a
summary document.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>